+++ CHANGELOG.md
@@ -7,5 +7,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- removed syscall_counter from ExecutionResources
 - Wrapped CallTraceV1 into Box for better performance
-